Revert "Simplify state, use only one reducer"
This reverts commit 0efaa95e3e8349b410106f4346e3f34c1cd68888.
