+++ src/components/StyleSwitch.js
@@ -69,10 +69,10 @@ StyleSwitch.propTypes = {
 
 const mapStateToProps = state => {
   return {
-    accessToken: state.mapboxAccessToken,
-    center: state.mapCoords.slice(0, 2),
-    mapStyle: state.mapStyle,
-    zoom: state.mapCoords[2]
+    accessToken: state.app.mapboxAccessToken,
+    center: state.app.mapCoords.slice(0, 2),
+    mapStyle: state.app.mapStyle,
+    zoom: state.app.mapCoords[2]
   };
 };
