+++ src/index.js
@@ -2,9 +2,10 @@ import React from "react";
 import ReactDOM from "react-dom";
 import { Provider } from "react-redux";
 import _ from "lodash";
-import { createStore, applyMiddleware, compose } from "redux";
-import { defaultState, reducer } from "./reducers/index";
+import { createStore, applyMiddleware, compose, combineReducers } from "redux";
+import { defaultState } from "./reducers/index";
 import apiCaller from "./middlewares/apiCaller";
+import reducers from "./reducers/index";
 import * as serviceWorker from "./serviceWorker";
 
 import App from "./components/App";
@@ -20,7 +21,9 @@ const initialState = _.merge({}, defaultState, persistedState);
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
 
 let store = createStore(
-  reducer,
+  combineReducers({
+    ...reducers
+  }),
   initialState,
   composeEnhancers(applyMiddleware(apiCaller))
 );
